+++ apps/client/src/shared/ui/pagination.tsx
@@ -1,7 +1,7 @@
 import { ChevronLeft, ChevronRight, MoreHorizontal } from "lucide-react";
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 import type { ButtonProps } from "@/shared/ui/button";
 import { buttonVariants } from "@/shared/ui/button";
